(feat): add databse seeder

+++ src/database/seeders/run-seeder.ts
@@ -1,35 +1,16 @@
-import mongoose from 'mongoose';
-import { logger } from 'src/utils';
-import config from '../../config';
+import { logger } from '../../utils';
 import usersDataSeed from './user.seed';
-import User from '@/modules/user/user.model';
+import User from '../../modules/user/user.model';
 
 const runSeeds = async () => {
-  mongoose
-    .connect(config.mongoose.url)
-    .then(() => {
-      logger.info('Connected to MongoDB');
-    })
-    .catch((err) => {
-      logger.info('connection error from seeder', err);
-    });
-
-  await User.deleteMany({});
-  await User.insertMany(usersDataSeed);
+  const count = await User.count({});
+  logger.info(`Users collection has ${count} datasets`);
+  if (count === 0) {
+    logger.info('running users seed...');
+    await User.deleteMany({});
+    await User.insertMany(usersDataSeed);
+    logger.info('users seeding completed!');
+  }
 };
 
-runSeeds().then(() => {
-  mongoose.connection.close();
-});
-
 export default runSeeds;
-
-// import { User } from '../../modules/user';
-// import usersDataSeed from './user.seed';
-
-// const runSeeds = async () => {
-//   await User.deleteMany({});
-//   await User.insertMany(usersDataSeed);
-// };
-
-// export default runSeeds;
